pyproject.toml
[project]
name = "011-brownieatelierresourcecheck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-common>=1.1.28",
    "azure-core>=1.34.0",
    "azure-identity>=1.23.0",
    "azure-mgmt-containerinstance>=10.1.0",
    "azure-mgmt-core>=1.5.0",
    "azure-mgmt-resource>=24.0.0",
    "python-decouple>=3.8",
]
